+++ CHANGELOG.md
@@ -7,6 +7,20 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+- [#15](https://github.com/nubank/emidje/pull/15): add the command
+  `emidje-run-all-tests-popup`, bound to `c-c c-j P`. This keybinding shows a
+  Magit popup that allow for users to set some switches/options to customize the
+  test execution.
+- [#15](https://github.com/nubank/emidje/pull/15): add ns filters to exclude
+  and/or include namespaces to be tested. Those options can be set through the
+  aforementioned popup.
+
+### Changed
+- `emidje-run-all-tests` (`c-c c-j p`) can't be called with a prefix argument
+  anymore to select test paths. Instead, users should use `c-c c-j P` to set
+  this option (and others) via a Magit popup menu.
+
 ## [1.1.0] - 2018-12-19
 
 ### Added